Proper transitioning, and restructure some
