+++ engine/src/main/java/games/aternos/odessa/engine/lobby/command/CreateArenaCommand.java
@@ -15,6 +15,9 @@ public class CreateArenaCommand implements CommandExecutor {
 
   @Override
   public boolean onCommand(CommandSender commandSender, Command command, String s, String[] strings) {
+    if(!this.gameLobbySystem.isActive()){
+      return true;
+    }
     if (!commandSender.isOp()) {
       return true;
     }
