+++ engine/src/main/java/games/aternos/odessa/engine/lobby/command/SetLobbyLocationCommand.java
@@ -20,7 +20,9 @@ public class SetLobbyLocationCommand implements CommandExecutor {
 
   @Override
   public boolean onCommand(CommandSender commandSender, Command command, String s, String[] strings) {
-
+    if(!this.gameLobbySystem.isActive()){
+      return true;
+    }
     if (!commandSender.isOp()) {
       return true;
     }
