+++ examples/hunger-games/src/main/java/games/aternos/odessa/example/hungergames/phase/ingame/InGamePhase.java
@@ -1,4 +1,4 @@
-package games.aternos.odessa.example.hungergames.phase;
+package games.aternos.odessa.example.hungergames.phase.ingame;
 
 import edu.umd.cs.findbugs.annotations.NonNull;
 import games.aternos.odessa.engine.service.player.PlayerService;
@@ -26,9 +26,9 @@ public class InGamePhase extends GamePhase {
    */
   @Override
   public void startPhase() {
-    Debug.$("test");
     this.setActive(true);
     playerService.dispersePlayers(this.getGame().getGameData().getPlayers(), this.getGame().getGameData().getGameArena());
+
   }
 
   @Override
