+++ examples/hunger-games/src/main/java/games/aternos/odessa/example/hungergames/phase/lobby/LobbyPhase.java
@@ -1,8 +1,9 @@
-package games.aternos.odessa.example.hungergames.phase;
+package games.aternos.odessa.example.hungergames.phase.lobby;
 
 import edu.umd.cs.findbugs.annotations.NonNull;
 import games.aternos.odessa.engine.lobby.GameLobbySystem;
-import games.aternos.odessa.example.hungergames.phase.runnable.LobbyPhaseRunnable;
+import games.aternos.odessa.example.hungergames.phase.ingame.InGamePhase;
+import games.aternos.odessa.example.hungergames.phase.lobby.runnable.LobbyPhaseRunnable;
 import games.aternos.odessa.gameapi.GameApi;
 import games.aternos.odessa.gameapi.game.Game;
 import games.aternos.odessa.gameapi.game.GameLifecycleManager;
@@ -15,11 +16,12 @@ public class LobbyPhase extends GamePhase {
 
   public LobbyPhase(@NonNull GameLifecycleManager owner, Game game) {
     super(owner, game);
+    this.setNextPhase(new InGamePhase(this.getOwner(), this.getGame()));
   }
 
   @Override
   public void initialize() {
-    this.setNextPhase(new InGamePhase(this.getOwner(), this.getGame()));
+
     this.setGamePhaseRunnable(new LobbyPhaseRunnable(this));
     gameLobbySystem = new GameLobbySystem(this.getOwner(), GameApi.getGameApi(), GameApi.getGameApi().getGame().getGameConfiguration());
   }
